backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
librosa
numpy
soundfile
python-dotenv
